// lib/api/generated/models/UpdateUserDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateUserDto = {
    /**
     * Nouveau nom d'utilisateur
     */
    username?: string;
    /**
     * Nouvelle adresse email
     */
    email?: string;
    /**
     * Nouveau mot de passe
     */
    password?: string;
    /**
     * URL du nouvel avatar
     */
    avatar?: string;
    /**
     * Statut publisher
     */
    isPublisher?: boolean;
};